Changes to CLI
Multiple changes to the argument parsing, including default values (There should no longer be any error cases, as the default values force security level 1 with 12 characters)

The error messages have been left in because it's still possible for a person to manually enter security level 0, or 5.

Seconly 2 new arguments have been added to print the password as stars in the console, and to omit logging the password.